% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bert_example_data}
\alias{bert_example_data}
\title{Example bert data}
\format{
A data frame with 4,035 rows and 1 column
\describe{
#' \item{message}{Text variable, the most common variable for \code{SegmentR} functions}
}
}
\usage{
bert_example_data
}
\description{
Example text data frame from an old social media API
}
\keyword{internal}
