% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{bertopic_detach}
\alias{bertopic_detach}
\title{Detach bertopic from the python session}
\usage{
bertopic_detach()
}
\value{
Nothing
}
\description{
Call this when you're finished with the topic modelling process. Although, safer may be to simply save your work and then restart your R session, as the Python session is still running (and as far as I know, there's no way to safely close)
}
