% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_models.R
\name{bt_base_ctfidf}
\alias{bt_base_ctfidf}
\title{Creare a base c-tf-idf model for skipping vectorisation step of bertopic pipeline}
\usage{
bt_base_ctfidf()
}
\value{
an empty c-tf-idf model (Python class)
}
\description{
Creare a base c-tf-idf model for skipping vectorisation step of bertopic pipeline
}
\examples{
base_ctf <- bt_base_ctfidf()

base_ctfidf <- bt_base_ctfidf()
}
