% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{bt_base_embedder}
\alias{bt_base_embedder}
\title{Create a base embedder for skipping embedding step of bertopic pipeline}
\usage{
bt_base_embedder()
}
\value{
an empty embedding model (Python class)
}
\description{
Create a base embedder for skipping embedding step of bertopic pipeline
}
\examples{
base_emebdder <- bt_base_embedder()

embedder <- bt_base_emebdder()
}
