% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_models.R
\name{bt_base_reducer}
\alias{bt_base_reducer}
\title{Create a base reducer for skipping dimensionality reduction step of bertopic pipeline}
\usage{
bt_base_reducer()
}
\value{
an empty dimensionality reduction model (Python class)
}
\description{
Create a base reducer for skipping dimensionality reduction step of bertopic pipeline
}
\examples{
base_reducer <- bt_base_reducer()

}
