% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bt_do_clustering}
\alias{bt_do_clustering}
\title{Cluster your data}
\usage{
bt_do_clustering(clustering_model, embeddings)
}
\arguments{
\item{clustering_model}{Python object, output of bt_make_clusterer*}

\item{embeddings}{Embeddings, output of bt_do_embedding or bt_do_reducing}
}
\value{
Cluster labels for each document
}
\description{
Cluster your data
}
\examples{
# create clustering model
clusterer <- bt_make_clusterer_kmeans(n_clusters = 2)

# mock embeddings
embeddings_test <-  matrix(runif(50), nrow = 25, ncol = 2)

# create clusters
clusters <- bt_do_clustering(clusterer, embeddings_test)

}
