% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{bt_empty_embedder}
\alias{bt_empty_embedder}
\title{Create an empty embedder for skipping embedding step of bertopic pipeline}
\usage{
bt_empty_embedder()
}
\value{
an empty embedding model (Python class)
}
\description{
Create an empty embedder for skipping embedding step of bertopic pipeline
}
\examples{
empty_emebdder <- bt_empty_embedder()

embedder <- bt_empty_embedder()
}
