% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_models.R
\name{bt_empty_reducer}
\alias{bt_empty_reducer}
\title{Create an empty reducer for skipping dimensionality reduction step of bertopic pipeline}
\usage{
bt_empty_reducer()
}
\value{
an empty dimensionality reduction model (Python class)
}
\description{
Create an empty reducer for skipping dimensionality reduction step of bertopic pipeline
}
\details{
This refers to the BaseDimensionalityReduction() function in the Python BERTopic library
and can be used to skip over the dimensionality reduction step that occurs as part of
fitting the bertopic model to the text data. This is convenient as it allows you to perform
dimensionality reduction of document embeddings externally to the model which gives full
transparency in terms of reduced embedding values and can save time if trialing fitting the model
with various parameters.
}
\examples{
empty_reducer <- bt_empty_reducer()

reducer <- bt_empty_reducer()
}
