% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{bt_fit_model}
\alias{bt_fit_model}
\title{Fit a topic model on your documents & embeddings}
\usage{
bt_fit_model(model, documents, embeddings, topic_labels = NULL)
}
\arguments{
\item{model}{Output of bt_compile_model() or another bertopic topic model}

\item{documents}{Your documents to topic model on}

\item{embeddings}{Your embeddings, can be reduced dimensionality or not}

\item{topic_labels}{Pre-existing labels, for supervised topic modelling}
}
\value{
a fitted BERTopic model
}
\description{
If you've already performed dimensionality reduction on your embeddings, you can feed in the reduced dimension embeddings to the embeddings argument, make sure to supply \code{bt_compile_model} with a base reducer (the output of \code{bt_base_reducer()})

NOTE: The bertopic model you are working with is a pointer to a python object
at a point in memory. This means that the input and the output model cannot be
differentiated between without explicitly saving the model before performing
this operation. A model is not returned as the function changes the input model.
}
