% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_transform_model.R
\name{bt_fit_transform_model}
\alias{bt_fit_transform_model}
\title{fit a bertopic model to cleaned data}
\usage{
bt_fit_transform_model(
  cleaned_text,
  calculated_embeddings = NULL,
  reducer = NULL,
  min_topic_size = 10,
  nr_topics = NULL,
  ngram_range = c(1, 1),
  embedding_model = "all-MiniLM-L6-v2",
  accelerator = "mps",
  diversity = 0.1,
  stopwords = TRUE,
  random_state = NULL
)
}
\arguments{
\item{cleaned_text}{cleaned text column that the model should be fit with}

\item{calculated_embeddings}{if embeddings are already calculated input embeddings as matrix}

\item{reducer}{the dimensionality reduction model to send to BERTopic()}

\item{min_topic_size}{minimum topic size}

\item{nr_topics}{the number of topics to find within the dataset}

\item{ngram_range}{ngram range for topic representation}

\item{embedding_model}{which embedding model to use or was used to produce embeddings}

\item{accelerator}{accelerator to use - default is mps, use NULL if none}

\item{diversity}{diversity of topic representation (1 = diverse, 0 = not diverse)}

\item{stopwords}{whether or not to remove stopwords in topic representations}

\item{random_state}{random state to pass to umap}
}
\value{
list containing fitted bertopic model and embeddings used to fit the model
}
\description{
fit a bertopic model to cleaned data
}
