% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bt_make_clusterer}
\alias{bt_make_clusterer}
\title{Create a clustering model}
\usage{
bt_make_clusterer(
  ...,
  clustering_method = c("hdbscan", "kmeans", "agglomerative", "base")
)
}
\arguments{
\item{...}{Arguments fed to the the clustering function determined by \verb{clustering_method =}}

\item{clustering_method}{A string defining which clustering method to use}
}
\value{
A clustering model (Python object)
}
\description{
Instantiates a clustering model which can be fed into a BertopicR pipeline.
}
\details{
Available clustering models are 'hdbscan', 'kmeans', 'agglomerative', and 'base'. For each of these options, see either the hdbscan, or sklearn official documentation for lists of arguments which can be fed to each clustering model. or type "?bt_make_clusterer_hdbscan", "?bt_make_clusterer_kmeans" etc. to see which named arguments have been supplied as a default for each clustering model.

Executing "bt_make_clusterer(clustering_method = 'base')" is the equivalent of calling `"bt_base_clusterer()" and returns an empty model which is used for streamlining the topic modelling process.
}
\examples{
empty_clustering_model <- bt_make_clusterer(clustering_method = "base")

hdbscan_model <- bt_make_clusterer(clustering_method = "hdbscan", min_cluster_size = 10L)

kmeans_model <- bt_make_clusterer(clustering_method = "kmeans", n_clusters = 10L)
}
