% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bt_make_clusterer_agglomerative}
\alias{bt_make_clusterer_agglomerative}
\title{Create an Agglomerative Clustering clustering model}
\usage{
bt_make_clusterer_agglomerative(n_clusters = 20L)
}
\arguments{
\item{n_clusters}{number of clusters to search for (enter as integer by typing L after the number)}
}
\value{
An Agglomerative Clustering clustering model (Python object)
}
\description{
Create an Agglomerative Clustering clustering model
}
\examples{
clustering_model <- bt_make_clusterer_agglomerative(15L)

agglomerative_model <- bt_make_clusterer_agglomerative(n_clusters = 10L)
}
