% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bt_make_clusterer_kmeans}
\alias{bt_make_clusterer_kmeans}
\title{Create a kmeans clustering model}
\usage{
bt_make_clusterer_kmeans(n_clusters = 10L)
}
\arguments{
\item{n_clusters}{number of clusters to search for (enter as integer by typing L after the number)}
}
\value{
A kMeans clustering model (Python object)
}
\description{
Create a kmeans clustering model
}
\examples{
clustering_model <- bt_make_clusterer_kmeans(15L)

kmeans_model <- bt_make_clusterer_kmeans(n_clusters = 10L)
}
