% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfidf.R
\name{bt_make_ctfidf}
\alias{bt_make_ctfidf}
\title{Create an instance of the ClassTfidfTransformer from the bertopic.vectorizers module}
\usage{
bt_make_ctfidf(reduce_frequent_words = TRUE, bm25_weighting = FALSE)
}
\arguments{
\item{reduce_frequent_words}{should frequent words be reduced? Default is TRUE.}

\item{bm25_weighting}{should BM25 weighting be used? Default is FALSE.}
}
\value{
A ctfidf model (Python object).
}
\description{
This function creates an instance of the ClassTfidfTransformer from the bertopic.vectorizers module,
with the provided arguments. It is used to generate representations for topics by selecting words which are frequent within a topic and less frequent in the entire corpus.
}
