% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{bt_make_embedder_flair}
\alias{bt_make_embedder_flair}
\title{Create an embedding model using a model available from the \href{https://flairnlp.github.io/}{Flair Library}}
\usage{
bt_make_embedder_flair(
  model,
  ...,
  flair_class = c("FlairEmbeddings", "TransformerWordEmbeddings",
    "TransformerDocumentEmbeddings", "WordEmbeddings")
)
}
\arguments{
\item{model}{name of model used to create embeddings}

\item{...}{Additional arguments to be passed to the selected Flair class}

\item{flair_class}{BertopicR is currently compatible with 4 Flair embedding classes: FlairEmbeddings, WordEmbeddings, TransformerWordEmbeddings and TransformerDocumentEmbeddings.
If you chose to perform word embeddings rather than document embeddings, bt_do_embedding will pool the word embeddings for each document to calculate a mean value.}
}
\value{
an embedding model, formed according to the model defined, that can be input to be_do_embedding to create document embeddings
}
\description{
Create an embedding model using a model available from the \href{https://flairnlp.github.io/}{Flair Library}
}
\examples{
\donttest{
# Flair Embedding, reducing chars_per_chunk to help with memory issues
embedder <- bt_make_embedder_flair(model = "news-forward", flair_class = "FlairEmbeddings", chars_per_chunk = 400L)

# Transformer Document Embedding
embedder <- bt_make_embedder_flair(model = "roberta-base", flair_class = "TransformerDocumentEmbeddings")
}
}
