% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{bt_make_embedder_hf}
\alias{bt_make_embedder_hf}
\title{Create an embedding model using the \href{https://huggingface.co/models}{hugging face library}}
\usage{
bt_make_embedder_hf(..., task = "feature-extraction", model = NULL)
}
\arguments{
\item{...}{Additional arguments sent to transformers.pipeline()}

\item{task}{The task defining the pipeline to be returned, this defaults to feature extraction.}

\item{model}{The model used by the pipeline to make predictions}
}
\value{
a pipeline formed according to the task defined
}
\description{
Create an embedding model using the \href{https://huggingface.co/models}{hugging face library}
}
\examples{
# define task and use default model for that task
embedder <- bt_make_embedder_hf(task = "feature-extraction")

# define model and use task specified for that model
embedder <- bt_make_embedder_hf(model = "distilbert-base-cased")

# define task and model
embedder <- bt_make_embedder_hf(task = "feature-extraction", model = "facebook/bart-base")

# define 
}
