% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{bt_make_embedder_openai}
\alias{bt_make_embedder_openai}
\title{Create an embedding model using a model available from the \href{https://platform.openai.com/docs/models/embeddings}{OpenAI Library}}
\usage{
bt_make_embedder_openai(
  model = "text-embedding-ada-002",
  openai_api_key = "sk-"
)
}
\arguments{
\item{model}{name of openai model used to create embeddings}

\item{openai_api_key}{OpenAI API key is required to use the OpenAI API and can be found on the OpenAI website}
}
\value{
an embedding model, formed according to the model defined, that can be input to be_do_embedding to create document embeddings
}
\description{
Create an embedding model using a model available from the \href{https://platform.openai.com/docs/models/embeddings}{OpenAI Library}
}
\examples{
\dontrun{
embedder <- bt_make_embedder_openai(model = "text-embedding-ada-002", openai_api_key = "sk-")
}
}
