% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{bt_make_reducer}
\alias{bt_make_reducer}
\title{Create dimensionality reduction model}
\usage{
bt_make_reducer(
  ...,
  n_neighbors = 15L,
  n_components = 5L,
  min_dist = 0,
  metric = "euclidean",
  random_state = 42L,
  verbose = TRUE
)
}
\arguments{
\item{...}{Sent to umap$UMAP for adding additional arugments}

\item{n_neighbors}{The size of local neighbourhood (in terms of number of neighboring data points) used
for manifold approximation (default: 15).}

\item{n_components}{The number of dimensions to reduce to (default: 5).}

\item{min_dist}{The minimum distance between points in the low-dimensional representation (default: 0.0).}

\item{metric}{The metric to use for distance computation (default: "euclidean").}

\item{random_state}{The seed used by the random number generator (default: 42).}

\item{verbose}{Logical flag indicating whether to report progress during the dimension reduction (default: TRUE).}
}
\value{
A matrix or data frame of the dimension-reduced data. The number of rows will be the same
as \code{embeddings}, and the number of columns will be \code{n_components}.
}
\description{
This function wraps the UMAP functionality from Python's umap-learn package for use in R via reticulate. It allows you to perform dimension reduction on high-dimensional data, its intended use is in a BertopicR pipeline/
}
\details{
If you're concerned about processing time, you most likely will only want to reduce the dimensions of your dataset once. In this case, you should call \code{reducer <- bt_base_reducer()}
}
