% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{bt_make_reducer_truncated_svd}
\alias{bt_make_reducer_truncated_svd}
\title{Created Truncated SVD dimensionality reduction model}
\usage{
bt_make_reducer_truncated_svd(
  ...,
  n_components,
  n_iter = 5,
  svd_solver = "randomized"
)
}
\arguments{
\item{...}{Sent to sklearn.decomposition Truncated SVD function for adding additional arguments}

\item{n_components}{Number of components to keep}

\item{n_iter}{Number of iterations for randomised svd solver. Not used if svd solver is "arpack".}

\item{svd_solver}{method for reducing components can be arpack or randomized}
}
\value{
Truncated SVD Model that can be input to bt_do_reducing to reduce dimensions of data
}
\description{
This function wraps the Truncated SVD (Single Value Decomposition) functionality
from Python's sklearn package for use in R via reticulate. It allows you to
perform dimension reduction on high-dimensional data.
Its intended use is in a BertopicR pipeline.
}
