% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_outliers_ctfidf}
\alias{bt_outliers_ctfidf}
\title{Redistributes outliers using c-TF-IDF scores}
\usage{
bt_outliers_ctfidf(model = model, documents, topics, threshold)
}
\arguments{
\item{model}{BERTopic model}

\item{documents}{documents to which the model was fit}

\item{topics}{current topics associated with the documents}

\item{threshold}{minimum probability for outlier to be reassigned}
}
\value{
df with document, old topic, new topic
}
\description{
Uses cosine similarity of document and topic c-TF-IDF to find the most
}
\details{
appropriate topic for each outlier document.
}
