% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_outliers_probs}
\alias{bt_outliers_probs}
\title{Redistributes outliers using probabilities}
\usage{
bt_outliers_probs(
  model = model,
  documents,
  topics,
  probability_matrix,
  threshold
)
}
\arguments{
\item{model}{BERTopic model}

\item{documents}{documents to which the model was fit}

\item{topics}{current topics associated with the documents}

\item{probability_matrix}{topic-document probability matrix (probability of each document being a member of each topic)}

\item{threshold}{minimum probability for outlier to be reassigned}
}
\value{
df with document, old topic, new topic
}
\description{
Uses HDBSCAN calculated probabilities to find the most appropriate topic for
}
\details{
each outlier document.
}
