% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_outliers_tokenset_similarity}
\alias{bt_outliers_tokenset_similarity}
\title{Redistributes outliers using tokenset c-TF-IDF scores}
\usage{
bt_outliers_tokenset_similarity(
  ...,
  model = model,
  documents,
  topics,
  window = 4,
  stride = 1,
  threshold
)
}
\arguments{
\item{...}{Optional or additional parameters passed to approximate_distribution function, e.g. batch_size}

\item{model}{BERTopic model}

\item{documents}{documents to which the model was fit}

\item{topics}{current topics associated with the documents}

\item{window}{size of the moving window which is the number of tokens in a tokenset}

\item{stride}{how far the window should move at each step (number of words to
skip when moving to next tokenset)}

\item{threshold}{minimum probability for outlier to be reassigned}
}
\value{
df with document, old topic, new topic
}
\description{
Divides documents into tokensets and calculates their c-TF-IDF similarity to
}
\details{
each topic. Similarities of each tokenset to each topic are summed to and the

outlier is redistributed to the most similar topic.
}
