% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{bt_representation_hf}
\alias{bt_representation_hf}
\title{Use Huggingface models to create topic representation}
\usage{
bt_representation_hf(
  task,
  hf_model,
  topic_model,
  default_prompt = "keywords",
  custom_prompt = NULL,
  nr_samples = 100,
  nr_repr_docs = 2,
  diversity = 10,
  ...
)
}
\arguments{
\item{task}{Task defining the pipeline that will be returned. See https://huggingface.co/transformers/v3.0.2/main_classes/pipelines.html for more information. Use "text-generation" for gpt-like models and "text2text-generation" for T5-like models}

\item{hf_model}{The model that will be used by the pipeline to make predictions}

\item{topic_model}{The fitted bertopic model}

\item{default_prompt}{Whether to use the "keywords" or "documents" default prompt. Passing a custom_prompt will render this argument NULL. Default is "keywords" prompt.}

\item{custom_prompt}{The custom prompt to be used in the pipeline. If not specified, the "keywords" or "documents" default_prompt will be used. Use "[KEYWORDS]" and "[DOCUMENTS]" in the prompt to decide where the keywords and documents are inserted.}

\item{nr_samples}{}

\item{nr_repr_docs}{}

\item{diversity}{}

\item{...}{}
}
\description{
Use Huggingface models to create topic representation
}
