% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{bt_representation_keybert}
\alias{bt_representation_keybert}
\title{Create representation model using keybert}
\usage{
bt_representation_keybert(
  top_n_words = 10,
  nr_docs = 50,
  nr_samples = 500,
  nr_candidate_words = 100,
  random_state = 42
)
}
\arguments{
\item{top_n_words}{number of keywords/phrases to be extracted}

\item{nr_docs}{number of representative docs to be examined for key words per cluster}

\item{nr_samples}{number of samples to select representative docs from for each cluster}

\item{nr_candidate_words}{number of words to examine per topic}

\item{random_state}{random state for sampling docs}
}
\value{
KeyBERTInspired representation model
}
\description{
This creates a representation model based on the KeyBERT algorithm. This model
can be used to update topic representation using the bt_update_topics function.
}
\details{
KeyBERT is a python package that is used for extraction of key words from
documents. This model could be paired with a vectoriser model to adjust the
ngram range of the key words / terms.
}
