% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{bt_representation_mmr}
\alias{bt_representation_mmr}
\title{Create representation model using Maximal Marginal Relevance}
\usage{
bt_representation_mmr(diversity = 0.1, top_n_words = 10)
}
\arguments{
\item{diversity}{How diverse representation words/phrases are. 0 = not diverse, 1 = completely diverse}

\item{top_n_words}{Number of keywords/phrases to be extracted}
}
\value{
MaximalMarginalRelevance representation model
}
\description{
Calculates the maximal marginal relevance between candidate words and documents.
Considers similarity between keywords and phrases and already selected keywords
and phrases and chooses representation based on this to maximise diversity.
}
