% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit a bertopic model to cleaned data}
\usage{
fit_model(
cleaned_text,
min_topic_size = NULL,
ngram_range = ,
embedding_model = "all-MiniLM-L6-v2",
accelerator = "mps",
diversity = 0.3,
stopwords = TRUE,
random_state = 42)
}
\arguments{
\item{cleaned_text}{cleaned text column that the model should be fit with}

\item{min_topic_size}{minimum topic size}

\item{ngram_range}{ngram range for topic representation}

\item{embedding_model}{which embedding model to use}

\item{diversity}{diversity of topic representation (1 = diverse, 0 = not diverse)}

\item{stopwords}{whether or not to remove stopwords in topic representations}

\item{random_state}{random state to pass to umap}

\item{use_mps}{accelerator to use - default is mps, use NULL if none}
}
\description{
fit a bertopic model to cleaned data
}
