% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{import_bertopic}
\alias{import_bertopic}
\title{Import the main Bertopic Python module}
\usage{
import_bertopic()
}
\value{
The imported 'bertopic' Python module.
}
\description{
Import the main Bertopic Python module
}
