% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{install_bertopic}
\alias{install_bertopic}
\title{Install BertopicR's Python Environment and Dependencies}
\usage{
install_bertopic()
}
\description{
This function installs Miniconda and creates a Conda
environment named "BertopicR" (or the value of the \code{BERTOPICR_ENV}
environment variable if user has set it/installed individually), and installs the required Python packages.
}
