% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedf.R
\name{makedf}
\alias{makedf}
\title{Title}
\usage{
makedf(
model
embeddings
original_text = df$message
cleaned_text = df$text_clean,
date = df$created_time,
sentiment = df$sentiment,
permalink = df$permalink)
}
\arguments{
\item{model}{bertopic model}

\item{embeddings}{embeddings used to generate the model}

\item{original_text}{uncleaned text column}

\item{cleaned_text}{cleaned text column which model was fitted on}

\item{date}{date column}

\item{sentiment}{sentiment column}

\item{permalink}{permalink column}
}
\value{
df with bertopic output merged with input columns from the sprinkr export
}
\description{
Title
}
