% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_topics.R
\name{visualize_topics}
\alias{visualize_topics}
\title{Visualize Topics using BERTopic}
\usage{
visualize_topics(model, filename = "intertopic_distance_map")
}
\arguments{
\item{model}{A BERTopic model object. The model must have the method \code{visualize_topics}.}

\item{filename}{A character string specifying the name of the HTML file to save the visualization.
The default value is "intertopic_distance_map". The filename should not contain illegal characters.}
}
\value{
The function does not return a value but saves an HTML file containing the visualization
and displays it in the current R environment.
}
\description{
This function visualizes the intertopic distance map of topics from a BERTopic model using Python's Plotly library.
The visualization is saved as an interactive HTML file, which can be opened and viewed in a web browser.
}
\examples{
\dontrun{
# Assuming 'topic_model' is a BERTopic model object
visualize_topics(model = topic_model, filename = "plot")
}
}
