% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_diff_terms.R
\name{viz_diff_terms}
\alias{viz_diff_terms}
\title{Function to find terms with the greatest difference between topics}
\usage{
viz_diff_terms(
merged_df,
text_var = text_clean,
topic_var = topic,
stopwords = TRUE,
hashtags = TRUE,
mentions = TRUE,
top_n = 15,
min_freq = 25,
include_outliers = FALSE,
type = c("lollipops", "bars")
plots = list(c(1,2), c(4,5)))
}
\arguments{
\item{merged_df}{output from makedf function.Can be any df that includes a topic column}

\item{text_var}{text diff_terms to be extracted from}

\item{topic_var}{column containing topic variable}

\item{stopwords}{remove stopwords?}

\item{hashtags}{remove hashtags?}

\item{mentions}{remove mentions?}

\item{top_n}{number of terms to extract}

\item{min_freq}{minimum number of times a term should appear for it to be considered}

\item{include_outliers}{include outlier (-1) bertopic category?}

\item{type}{lollipop or bar chart}

\item{plots}{specific plots to output. Should be input as c(x, y) or if you want more than 1
chart list(c(x, y), c(u, v)) where x, y, u and v are topic numbers.}
}
\value{
a ggplot object of top different terms between each pair of topics
}
\description{
Function to find terms with the greatest difference between topics
}
