% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_top_terms.R
\name{viz_top_terms}
\alias{viz_top_terms}
\title{create top terms charts for topic modelling completed using bertopic}
\usage{
viz_top_terms(
merged_df,
text_var = text_clean,
 topic_var = topic,
stopwords = TRUE,
hashtags = TRUE,
mentions = TRUE,
top_n = 15,
n_row = 2,
min_freq = 25,
include_outliers = FALSE,
type = "lollipops")
}
\arguments{
\item{merged_df}{df output from makedf function.Can be any df that includes a topic column}

\item{text_var}{text top terms to be extracted from}

\item{topic_var}{column containing topic variable}

\item{stopwords}{remove (english) stopwords?}

\item{hashtags}{remove hashtags?}

\item{mentions}{remove mentions?}

\item{top_n}{number of terms to extract}

\item{n_row}{number of rows the plots should take up}

\item{min_freq}{minimum number of times a term should appear for it to be considered}

\item{include_outliers}{include the -1 (outlier) bertopic category}

\item{type}{lollipop or bar chart?}
}
\value{
list of all_terms and max_only top term bar charts
}
\description{
create top terms charts for topic modelling completed using bertopic
}
