% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{bt_compile_model}
\alias{bt_compile_model}
\title{Build a BERTopic model}
\usage{
bt_compile_model(
  embedding_model = NULL,
  reduction_model = NULL,
  clustering_model = NULL,
  vectoriser_model = NULL,
  ctfidf_model = NULL
)
}
\arguments{
\item{embedding_model}{Model for creating embeddings (Python object)}

\item{reduction_model}{Model for reducing embeddings' dimensions (Python object)}

\item{clustering_model}{Model for clustering (Python object)}

\item{vectoriser_model}{Model for vectorising input for topic representations (Python object)}

\item{ctfidf_model}{Model for performing class-based tf-idf (ctf-idf) (Python object)}
}
\value{
a BERTopic model
}
\description{
Keep \verb{*_model} = NULL to proceed with a model made from default parameters (see each individual \verb{make_*} function for parameters). However, it is not advisable to accept default parameters for each model; you should tune each model according to your dataset and the business question you are answering.
}
