% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{bt_do_reducing}
\alias{bt_do_reducing}
\title{Perform dimensionality reduction on your embeddings}
\usage{
bt_do_reducing(reducer, embeddings)
}
\arguments{
\item{reducer}{Your dimensionality reduction model}

\item{embeddings}{Your embeddings}
}
\value{
Embeddingd with reduced number of dimensions
}
\description{
Perform dimensionality reduction on your embeddings
}
\examples{
reducer <- bt_make_reducer_umap(n_neighbours = 2, n_components = 2)
embeddings <- matrix(runif(50), nrow = 5, ncol = 10)
reduced_embeddings <- bt_do_reducing(reducer, embeddings)
}
