% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducer.R
\name{bt_do_reducing}
\alias{bt_do_reducing}
\title{Perform dimensionality reduction on your embeddings}
\usage{
bt_do_reducing(reducer, embeddings)
}
\arguments{
\item{reducer}{Your dimensionality reduction model}

\item{embeddings}{Your embeddings}
}
\value{
Embeddingd with reduced number of dimensions
}
\description{
Perform dimensionality reduction on your embeddings
}
