% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{bt_empty_clusterer}
\alias{bt_empty_clusterer}
\title{Create an empty clusterer for skipping clustering step of bertopic pipeline}
\usage{
bt_empty_clusterer()
}
\value{
an empty clustering model (Python class)
}
\description{
Create an empty clusterer for skipping clustering step of bertopic pipeline
}
\examples{
empty_clusterer <- bt_empty_clusterer()

clusterer <- bt_empty_clusterer()
}
