% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_models.R
\name{bt_empty_embedder}
\alias{bt_empty_embedder}
\title{Create an empty embedder for skipping embedding step of bertopic pipeline}
\usage{
bt_empty_embedder()
}
\value{
an empty embedding model (Python class)
}
\description{
Create an empty embedder for skipping embedding step of bertopic pipeline
}
\details{
This refers to the BaseEmbedder() function in the Python BERTopic library
and can be used to skip over the document embedding step that occurs as part of
fitting the bertopic model to the text data. This is convenient as it allows you to perform
document embedding externally to the model which gives full
transparency in terms of embedding values and can save time if trialing fitting the model
with various parameters.
}
\examples{
empty_emebdder <- bt_empty_embedder()

embedder <- bt_empty_embedder()
}
