% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{bt_empty_reducer}
\alias{bt_empty_reducer}
\title{Create an empty reducer for skipping dimensionality reduction step of bertopic pipeline}
\usage{
bt_empty_reducer()
}
\value{
an empty dimensionality reduction model (Python class)
}
\description{
Create an empty reducer for skipping dimensionality reduction step of bertopic pipeline
}
\examples{
empty_reducer <- bt_empty_reducer()

reducer <- bt_empty_reducer()
}
