% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{bt_fit_model}
\alias{bt_fit_model}
\title{Fit a topic model on your documents & embeddings}
\usage{
bt_fit_model(model, documents, embeddings, topic_labels = NULL)
}
\arguments{
\item{model}{Output of bt_compile_model() or another bertopic topic model}

\item{documents}{Your documents to topic model on}

\item{reduced_embeddings}{Dimensionality reduced embeddings}
}
\value{
a fitted BERTopic model
}
\description{
Fit a topic model on your documents & embeddings
}
