% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{bt_make_clusterer_hdbscan}
\alias{bt_make_clusterer_hdbscan}
\title{Create an HDBSCAN clustering model}
\usage{
bt_make_clusterer_hdbscan(
  ...,
  min_cluster_size = 10L,
  min_samples = 10L,
  metric = "euclidean",
  cluster_selection_method = "eom",
  prediction_data = FALSE
)
}
\arguments{
\item{...}{Additional arguments sent to hdbscan.HDBSCAN()}

\item{min_cluster_size}{Minimum number of data points for each cluster, enter as integer by adding L to number}

\item{metric}{Distance metric to calculate clusters with}

\item{cluster_selection_method}{The method used to select clusters. Default is "eom".}

\item{prediction_data}{Logical}
}
\value{
An instance of the HDBSCAN clustering model (Python object.
}
\description{
Instantiates an HDBSCAN clustering model using the hdbscan Python library.
}
\examples{
clustering_model <- bt_make_clusterer_hdbscan(metric = "minkowski")
}
\seealso{
\url{https://hdbscan.readthedocs.io/en/latest/}
\url{https://hdbscan.readthedocs.io/en/latest/how_hdbscan_works.html}
\url{https://hdbscan.readthedocs.io/en/latest/basic_hdbscan.html#what-about-different-metrics}
}
