% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{bt_make_embedder_spacy}
\alias{bt_make_embedder_spacy}
\title{Create an embedding model using a model available from the \href{https://spacy.io/models}{Spacy Library}}
\usage{
bt_make_embedder_spacy(model, ..., exclude = NULL)
}
\arguments{
\item{model}{The pipeline used to make predictions}

\item{...}{additional arguments to be sent to the spacy.load function}

\item{exclude}{name of pipeline components to exclude}
}
\value{
a pipeline formed according to the model defined
}
\description{
Create an embedding model using a model available from the \href{https://spacy.io/models}{Spacy Library}
}
\examples{
# specify a non-transformer model, excluding features not required
embedder <- bt_make_embedder_spacy(model = "en_core_web_md", exclude = c("tagger", "parser", "ner", "attribute_ruler", "lemmatizer"))

# specify a transformer model and exclude features not required
embedder <- bt_make_embedder_spacy(model = "en_core_web_trf", exclude = c("tagger", "parser", "ner", "attribute_ruler", "lemmatizer"))

}
