% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{bt_make_reducer_pca}
\alias{bt_make_reducer_pca}
\title{Create pca dimensionality reduction model}
\usage{
bt_make_reducer_pca(..., n_components, svd_solver = "auto")
}
\arguments{
\item{...}{Sent to sklearn.decomposition UMAP function for adding additional arguments}

\item{n_components}{Number of components to keep}

\item{svd_solver}{method for reducing components can be auto, full, arpack, randomized}
}
\value{
A PCA Model that can be input to bt_do_reducing to reduce dimensions of data
}
\description{
This function wraps the PCA functionality from Python's sklearn package for
use in R via reticulate. It allows you to perform dimension reduction on
high-dimensional data, its intended use is in a BertopicR pipeline.
}
