% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{bt_make_reducer_truncated_svd}
\alias{bt_make_reducer_truncated_svd}
\title{Created Truncated SVD dimensionality reduction model}
\usage{
bt_make_reducer_truncated_svd(
  n_components,
  ...,
  n_iter = 5L,
  svd_solver = c("randomized", "arpack")
)
}
\arguments{
\item{n_components}{Number of components to keep}

\item{...}{Sent to sklearn.decomposition Truncated SVD function for adding additional arguments}

\item{n_iter}{Number of iterations for randomised svd solver. Not used if svd solver is "arpack".}

\item{svd_solver}{method for reducing components can be arpack or randomized}
}
\value{
Truncated SVD Model that can be input to bt_do_reducing to reduce dimensions of data
}
\description{
This function wraps the Truncated SVD (Single Value Decomposition) functionality
from Python's sklearn package for use in R via reticulate. It allows you to
perform dimension reduction on high-dimensional data.
Its intended use is in a BertopicR pipeline. If you're concerned about processing
time, you most likely will only want to reduce the dimensions of your dataset once.
In this case, when compiling your model with bt_compile_model you should call
\code{reducer <- bt_empty_reducer()}.
}
\examples{
reducer <- bt_make_reducer_truncated_svd(n_components = 5)
reducer <- bt_make_reducer_truncated_svd(n_components = 5, svd_solver = "randomized", n_iter = 10)
}
\seealso{
\url{https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.TruncatedSVD.html}
\url{https://scikit-learn.org/stable/modules/decomposition.html#lsa}
}
