% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_merge_topics}
\alias{bt_merge_topics}
\title{Merge topics
Merge topics that have been generated with a BERTopic model}
\usage{
bt_merge_topics(model = model, documents, topics_to_merge = list())
}
\arguments{
\item{model}{Output of bt_compile_model() or another bertopic topic model}

\item{documents}{Your documents to topic model on}

\item{topics_to_merge}{list of topics created by your bertopic model that you wish to merge}
}
\value{
bertopic model with specified topics merged
}
\description{
Merge topics
Merge topics that have been generated with a BERTopic model
}
