% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_outliers_ctfidf}
\alias{bt_outliers_ctfidf}
\title{Redistributes outliers using c-TF-IDF scores}
\usage{
bt_outliers_ctfidf(fitted_model, documents, topics, threshold = 0.3)
}
\arguments{
\item{fitted_model}{Output of bt_fit_model() or another bertopic topic model. The model must have been fitted to data.}

\item{documents}{documents to which the model was fit}

\item{topics}{current topics associated with the documents}

\item{threshold}{minimum probability for outlier to be reassigned}
}
\value{
df with document, old topic, new topic
}
\description{
Calculates the cosine similarity of c-TF-IDF between documents
and topics and redistributes outliers based on the topic it has the highest
similarity to. Note that the purpose
of this function is to obtain a new list of topics that can then be used to
update the model, it does not make any changes to the model itself, the topic
classification the model outputs does not change after running this function.
The bt_update_topics function needs to be used to make the change to the model
itself.
}
