% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_outliers_embeddings}
\alias{bt_outliers_embeddings}
\title{Redistributes outliers using embeddings}
\usage{
bt_outliers_embeddings(model = model, documents, topics, embeddings, threshold)
}
\arguments{
\item{model}{BERTopic model}

\item{documents}{documents to which the model was fit}

\item{topics}{current topics associated with the documents}

\item{embeddings}{embeddings used to create topics}

\item{threshold}{minimum probability for outlier to be reassigned}
}
\value{
df with document, old topic, new topic
}
\description{
Uses cosine similarity of embeddings to find the most appropriate topic for
}
\details{
each outlier document.
}
