% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_topics.R
\name{bt_probability_matrix}
\alias{bt_probability_matrix}
\title{Calculates Topic-Document Probability Matrix}
\usage{
bt_probability_matrix(fitted_model)
}
\arguments{
\item{fitted_model}{Output of bt_fit_model() or another bertopic topic model. The model must have been fitted to data.}
}
\value{
topic-document probability matrix
}
\description{
Uses hdbscan soft clustering to create a matrix with a row per document and
column per model topic. The intersection of row and column is the probability
that that document is part of that topic. This is only applicable for models
created using a hdbscan clustering method (bt_make_clusterer_hdbscan or equivalent)
}
