% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation.R
\name{bt_representation_openai}
\alias{bt_representation_openai}
\title{Create representation model that uses OpenAI text generation models}
\usage{
bt_representation_openai(
fitted_model,
documents,
openai_model = "gpt-4o-mini",
nr_repr_docs = 10,
nr_samples = 500,
chat = TRUE,
api_key = "sk-",
delay_in_seconds = NULL,
prompt = NULL,
diversity = NULL)
}
\arguments{
\item{fitted_model}{Output of bt_fit_model() or another bertopic topic model. The model must have been fitted to data.}

\item{documents}{documents used to fit the fitted_model}

\item{openai_model}{openai model to use. If using a chat model, set chat = TRUE}

\item{nr_repr_docs}{number of representative documents per topic to send to the openai model}

\item{nr_samples}{Number of sample documents from which the representative docs are chosen}

\item{chat}{set to TRUE if using gpt-3.5 model}

\item{api_key}{OpenAI API key is required to use the OpenAI API and can be found on the OpenAI website}

\item{delay_in_seconds}{The delay in seconds between consecutive prompts, this is to avoid rate limit errors.}

\item{prompt}{The prompt to be used with the openai model. If NULL, the default prompt is used.}

\item{diversity}{diversity of documents to be sent to the huggingface model. 0 = no diversity, 1 = max diversity.}
}
\value{
OpenAI representation model
}
\description{
Representative documents are chosen from each topic by sampling (nr_samples)
a number of documents from the topic and calculating which of those documents are
most representative of the topic by c-tf-idf cosine similarity between the topic and the
individual documents. From this the most representative documents (the number
is defined by the nr_repr_docs parameter) is extracted and passed to  the OpenAI API
to generate topic labels based on one of their Completion (chat = FALSE) or ChatCompletion
(chat = TRUE) models.
}
