% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{install_bertopic}
\alias{install_bertopic}
\title{Install BertopicR's Python Environment and Dependencies}
\usage{
install_bertopic()
}
\description{
This function ensures Miniconda is available and then sets up the necessary
conda environment using the package's \code{environment.yml} file.
}
\details{
After running this function, you \strong{must} restart your R session for the
changes to take effect.
